==> r-code/01_rawdata.Rmd <==
---
title: "Creation of rawdata"
author: Bernd Panassiti
output: html_document
date: "2024-11-28"
editor_options: 
  chunk_output_type: console
---


```{r setup, include=FALSE}
library(knitr)
opts_knit$set(root.dir='../')                     # definining working directory; or normalizePath('../')
opts_chunk$set(fig.align='center',                # aligns all figures
                echo=TRUE,                        # shows r-code
                message=FALSE,                    # suppresses library outputs
                warnings=FALSE,                   # suppresses library outputs
                tidy=TRUE,  # prevents the source code from running off a pdf page
                dev='pdf')
```

```{r}
#Install/load pacman
if(!require(pacman)){install.packages("pacman");require(pacman)}
#Install packages
p_load("readxl")
```

# Vegetation

veg_layer_red (ohne cf. Arten und Alchemilla pratensis zu A. pratensis agg. hinzugefügt, Viola sp. raus) 

```{r}
GermanSl_dbf <- foreign::read.dbf("data/rawdata/ecodbase.dbf")
veg_xlsx <- readxl::read_excel("data/rawdata/veg.xlsx")
veg_layer_red_xlsx <- readxl::read_excel("data/rawdata/veg_layer_red.xlsx")
vegTreatment_xlsx  <- readxl::read_excel("data/rawdata/veg_treatment.xlsx")
```

# treatments
```{r}
treatment_xlsx <- read_excel("data/rawdata/treatment_neu.xlsx")
```

#Plot Data
```{r}
plot_data_xlsx <- readxl::read_excel("data/rawdata/Plot_Data.xlsx",
                                     col_types = c("text", "numeric", "numeric", 
         "numeric", "text", "text", "numeric", 
         "numeric", "numeric", "numeric", 
         "numeric", "numeric", "numeric", 
         "numeric", "numeric", "numeric", 
         "numeric", "numeric", "numeric", 
         "numeric", "numeric", "numeric", 
         "numeric", "numeric", "numeric", 
         "numeric", "numeric", "numeric", 
         "numeric", "numeric", "numeric", 
         "numeric", "numeric", "numeric", 
         "numeric", "date", "numeric", "numeric", 
         "numeric", "numeric"), na = "NA")

coords_xlsx = readxl::read_excel("data/rawdata/Koordinaten_Plots.xlsx")

```


#Zeigerwerte
```{r}
zeigerwerte_xlsx <- read_excel("data/rawdata/Zeigerwerte_ergaenzt.xlsx")
zeigerwerte_ergaenzt_xlsx <- readxl::read_excel("data/rawdata/Zeigerwerte_ergaenzt.xlsx")


zeigerwerte_mittel_xlsx <- read_excel("data/rawdata/Zeigerwerte_mittel_hl_sl.xlsx")

lichtzahlen_xlsx <- read_excel("data/rawdata/Lichtzahlen.xlsx")
temperaturzahlen_xlsx <- read_excel("data/rawdata/Temperaturzahlen.xlsx")
kontizahlen_xlsx <- read_excel("data/rawdata/Kontinentalitaetszahlen.xlsx")
feuchtezahlen_xlsx <- read_excel("data/rawdata/Feuchtezahlen.xlsx")
reaktionszahlen_xlsx <- read_excel("data/rawdata/Reaktionszahlen.xlsx")
naehrstoffzahlen_xlsx <- read_excel("data/rawdata/Naehrstoffzahlen.xlsx")
salzzahlen_xlsx <- read_excel("data/rawdata/Salzzahlen.xlsx")
temperaturzahlen_xlsx <- readxl::read_excel("data/rawdata/Temperaturzahlen.xlsx")
```


# Forest association
https://www.nw-fva.de/forschen/projekte/waldartenlisten

Innerhalb der Waldartenliste werden vier Gruppen der Waldbindung unterschieden, die zwei
Hauptgruppen zugeordnet sind:
1 weitgehend an Wald gebunden
1.1 vorwiegend im geschlossenen Wald
1.2 vorwiegend an Waldrändern und auf Waldverlichtungen
2 im Wald und im Offenland
2.1 im Wald wie im Offenland
2.2 auch im Wald, aber Schwerpunkt im Offenland

Eine weitere Differenzierung erfolgt über die vorangestellten Buchstaben. Bei Gefäßpflanzen werden die
Buchstaben B (Bäume), S (Sträucher), K (Kräuter) sowie E (Epiphyten), bei Moosen M und bei Flechten
F verwendet.

Hauptgruppe 1 enthält alle weitgehend an Wald gebundenen Taxa, während die in Hauptgruppe 2
zusammengefassten sowohl im Wald als auch im Offenland vorkommen. Die Arten mit dem höchsten
Grad der Waldbindung sind in Hauptgruppe 1 zusammengefasst. Gruppe 1.1 beinhaltet diejenigen Taxa,
die vorwiegend im geschlossenen Wald zu finden sind, während Gruppe 1.2 vorwiegend für Waldränder
und Waldverlichtungen typische Sippen umfasst. Hierzu zählen Arten der Waldaußen- und
Waldinnensäume ebenso wie Arten, die nach Windwurf bzw. Kahlschlag oder durch die Anlage von
Rückelinien und Wegen gefördert werden. Auch viele Arten mit enger Bindung an standortbedingt lichte
Wälder, vor allem auf flachgründigen oder blockreichen Hängen, gehören in die Gruppe 1.2. Nahezu alle
in dieser Gruppe genannten Sippen können auch im geschlossenen Wald vorkommen, sind dort aber meist
nur mit geringen Mengenanteilen zu finden. Ein Teil der in Hauptgruppe 1 eingestuften Arten kann
vereinzelt auch in von Gehölzen geprägten Lebensräumen außerhalb des Waldes auftreten, so
beispielsweise im Bereich von Hecken, Feldgehölzen, Latschen- und Grünerlengebüschen.
Arten der Hauptgruppe 2 sind sowohl im Wald als auch im Offenland verbreitet. Die Gruppe 2.1 enthält
Arten, die im Wald gleichermaßen wie im Offenland zu finden sind. Die Arten der Gruppe 2.2 haben
dagegen ihren Schwerpunkt im Offenland, kommen aber auch regelmäßig im Wald vor. Für eine
Zuordnung zu einer der beiden Gruppen spielt die aktuelle Häufigkeit der einzelnen Arten im Wald bzw.
im Offenland keine Rolle. Entscheidend ist in Zweifelsfällen die Frage, ob eine Art Bestandteil naturnaher
Waldtypen ist (Gruppe 2.1) oder nicht (Gruppe (2.2).
Ein großer Teil der Arten von Gruppe 2.1 hat seine ursprünglichen Wuchsorte vermutlich in Wäldern und
ist von hier in die Offenland-Lebensräume der Kulturlandschaft eingewandert bzw. konnte nach Rodung
der Wälder am Wuchsort überdauern. Dagegen zählen zur Gruppe 2.2 viele Sippen, die erst durch Aufforstung
von Offenland (Zwergstrauchheiden, Kalk- und Sandmagerrasen, Feuchtwiesen etc.) zu Bewohnern
lichter Wälder wurden. Hierher gehören auch Arten beweideter Waldstandorte, wie sie heute fast
nur noch in den Alpen vorkommen. Insgesamt zeichnen sich viele Arten der Gruppe 2.2 durch einen
hohen Lichtbedarf aus (Abb. 2). Da die Arten der Gruppe 2.2 in einigen Waldtypen, z. B. in Kiefernwäldern,
als „Sukzessionsreste“ über lange Zeiträume existieren und hier charakteristisch sind, werden sie in
der Liste der Waldgefäßpflanzen aufgeführt, auch wenn ihr Schwerpunkt im Offenland liegt.


```{r}
waldarten_xlsx <- read_excel("data/rawdata/Waldarten.xlsx")
```

# Red list
```{r}
rl_bayern_xlsx <- read_excel("data/rawdata/RL_Bayern.xlsx")
```

# bioflor, phytosociology, biotop $30

Bestimmungsschlüssel für Flächen nach §30 BNatSchG/Art. 23 BayNatSchGStand

Tafel 33: Grünland mittlerer bis höherer Gebirgslagen
Kriterien für Tafel 33
Die Gesamtdeckung der Arten des obenstehenden Blocks muss
mindestens 50 % betragen (A ≥ 4).
Es müssen mindestens eine der mit 1 bezeichneten Arten, zwei der mit 2 bezeichneten, drei der mit 2 oder 3 bezeichneten Arten oder vier der mit 2, 3 oder 4 bezeichneten Arten vorkommen.


Tafel 34: Grünland mäßig frischer bis trockener Standorte (von Tafel 9, 30 oder 41)
Kriterien für Tafel 34
1. Die Gesamtdeckung der unten aufgeführten Arten muss mindestens 25 % betragen (A ≥ 3).*
stehenden Blocks sind nicht
2. Es müssen mindestens eine der mit 1 bezeichneten Arten, zwei der mit 2 bezeichneten, drei der mit 2 oder 3 bezeichneten Arten oder vier der mit 2, 3 oder 4 bezeichneten Arten vorkommen.


Tafel 35: Grünlandbestände auf Mager- oder Trockenstandorten mittlerer und tieferer Lagen
Trockenrasen, Halbtrockenrasen, Sandmagerrasen, Borstgrasrasen und andere Magerrasengesellschaften in der Regel zweimal jährlich
(extensiv beweidet bzw. i.d.R. nur einmal jährlich gemäht, inkl. junge magere Brachen)

Kriterien für Tafel 35
Es müssen mindestens eine der mit 1 bezeichneten Arten, zwei der mit 2 bezeichneten oder 3 der mit 2 oder 3 bezeichneten Arten vorkommen.

```{r}
bioflorBiotop_xlsx <- read_excel("data/rawdata/BiolFlor_biotop.xlsx",
                                  na = "NA")
```


# save
```{r}
save(
  plot_data_xlsx,coords_xlsx,
  veg_xlsx,veg_layer_red_xlsx,
  vegTreatment_xlsx,
  GermanSl_dbf,
  treatment_xlsx,
  #indicator values
  zeigerwerte_xlsx,zeigerwerte_ergaenzt_xlsx,
  zeigerwerte_mittel_xlsx,
  
  lichtzahlen_xlsx,
  temperaturzahlen_xlsx,
  kontizahlen_xlsx,
  feuchtezahlen_xlsx,
  reaktionszahlen_xlsx,
  naehrstoffzahlen_xlsx,
  salzzahlen_xlsx,
  temperaturzahlen_xlsx,
  
  # forest association, RL
  waldarten_xlsx,rl_bayern_xlsx,
  bioflorBiotop_xlsx,
  file="data/rawdata.RData")
```
