==> r-code/02_workingdata.Rmd <==
---
title: "Preparation of workingdata"
author: "Bernd Panassiti"
date: "2024-11-28"
output:
  pdf_document: default
  html_document: default
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
library(knitr)
opts_knit$set(root.dir='../')                     # definining working directory; or normalizePath('../')
opts_chunk$set(fig.align='center',                # aligns all figures
                echo=TRUE,                        # shows r-code
                message=FALSE,                    # suppresses library outputs
                warnings=FALSE,                   # suppresses library outputs
                tidy=TRUE,  # prevents the source code from running off a pdf page
                dev='pdf')
```

```{r}
#Install/load pacman
if(!require(pacman)){install.packages("pacman");require(pacman)}
#Install packages
p_load(ggplot2,tidyverse,dplyr,magrittr)
```

# load raw data

```{r}
load("data/rawdata.RData")
```

# save rawdata as new r-object
```{r}
veg           = veg_xlsx
veg_layer_red = veg_layer_red_xlsx
veg           = as.data.frame(veg)
vegTreatment  = vegTreatment_xlsx
treatment     = treatment_xlsx
treatment     = as.data.frame(treatment)
plot_data     = plot_data_xlsx
plot_data     = as.data.frame(plot_data)
zeigerwerte   = zeigerwerte_xlsx
zeigerwerte   = as.data.frame(zeigerwerte)
zeigerwerte_mittel = zeigerwerte_mittel_xlsx

lichtzahlen       = lichtzahlen_xlsx
temperaturzahlen  = temperaturzahlen_xlsx
kontizahlen       = kontizahlen_xlsx
feuchtezahlen     = feuchtezahlen_xlsx
reaktionszahlen   = reaktionszahlen_xlsx
naehrstoffzahlen  = naehrstoffzahlen_xlsx
salzzahlen        = salzzahlen_xlsx
temperaturzahlen  = temperaturzahlen_xlsx

GermanSl      = GermanSl_dbf
coords        = coords_xlsx

  # forest association, RL, biotop species etc.
waldarten      = as.data.frame(waldarten_xlsx)
rl_bayern      = as.data.frame(rl_bayern_xlsx)
bioflorBiotop  = as.data.frame( bioflorBiotop_xlsx)
```

```{r}
plot_data$Alm = as.factor(plot_data$Alm)
plot_data$Alm_ID = as.factor(plot_data$Alm_ID)
plot_data$Transekt = as.factor(plot_data$Transekt)
plot_data <-  plot_data[order(plot_data$Plot_ID), ]
rownames(plot_data) <- NULL
# sorted according height
plot_data$Alm <- factor(plot_data$Alm , levels = c("Saletalm", "Engertalm",
                                                   "Bindalm","Königsbachalm",
                                                   "Halsalm","Königsbachalm 2",
                                                   "Mittereisalm","Königsbergalm"))
plot_data$Elevation <- round(plot_data$Elevation, 1)
plot_data$Verbuschung <- as.factor(plot_data$Verbuschung)
```

# Subset of herblayer
Subset nur mit Herblayer und Seedling Layer
```{r}
veg_layer <- subset(veg, layer %in% c("hl", "sl"))
veg_layer <- data.frame(veg_layer, row.names = NULL)
rownames(veg_layer) <- veg_layer[,1]
veg_layer <- veg_layer[, -c(1, 2)]

veg_layer <- t(veg_layer)
veg_layer <- as.data.frame(veg_layer)
veg_layer <-  veg_layer[order(rownames(veg_layer)), ]

# veg_layer_red (ohne cf. Arten und Alchemilla pratensis zu A. pratensis agg. hinzugefügt, Viola sp. raus) 
veg_layer_red <- data.frame(veg_layer_red, row.names = NULL)
rownames(veg_layer_red) <- veg_layer_red[,1]
veg_layer_red <- veg_layer_red[, -c(1, 2)]

veg_layer_red <- t(veg_layer_red)
veg_layer_red <- as.data.frame(veg_layer_red)
veg_layer_red <-  veg_layer_red[order(rownames(veg_layer_red)), ]
```


#Zeigerwerte auf Vegetationstabelle anwenden

In der Tabelle ecodbase fehlen für zahlreiche Arten die Zeigerwerte, wahrscheinlich wegen taxonomischer Differenzen zwischen der Artenliste von GermanSl und Ellenberg 1991. Darum wurden die fehlenden Zeigerwerte mit den Werten von Ellenberg & Leuschner (2010) ergänzt.

```{r Zeigerwerte}
names(GermanSl)[names(GermanSl) == "ABBREVIAT"] <- "species"
veg_ellenberg <- dplyr::left_join(veg, GermanSl, by = "species")
print(veg_ellenberg)


veg_ellenberg_hl <- subset(veg_ellenberg, layer %in% c("hl")) ### Wählt nur Zeilen mit layer = hl
rownames(veg_ellenberg_hl) <- veg_ellenberg_hl[,1] ### Artnamen als Zeilennamen
veg_ellenberg_hl <- veg_ellenberg_hl[, -c(1, 2)] ### Artnamenspalte und Layerspalte löschen
veg_selected <- veg_ellenberg_hl[, 1:144] ### Nur ausgewälte Daten für NMDS
Zeigerwerte <- veg_ellenberg_hl[, c("OEK_L", "OEK_T", "OEK_K", "OEK_F", "OEK_R", "OEK_N", "OEK_S")]
Zeigerwerte_ohneNA <- na.omit(Zeigerwerte)



rownames(zeigerwerte) <- zeigerwerte[,1]
zeigerwerte_withoutSpeciesColumn <- zeigerwerte[, -c(1)]
zeigerwerte_transposed <- as.data.frame(t(zeigerwerte_withoutSpeciesColumn))
zeigerwerte_transposed_rownames <- cbind(Name = rownames(zeigerwerte_transposed), zeigerwerte_transposed)

# Versuch mit veg_layer, also nur mit hl und sl!

# Subset für NMDS vorbereiten
veg_nmds <- subset(veg, layer %in% c("hl", "sl"))
veg_nmds <- data.frame(veg_nmds, row.names = NULL)
rownames(veg_nmds) <- veg_nmds[,1]
veg_nmds <- veg_nmds[, -c(1, 2)]

veg_nmds <- t(veg_nmds)
veg_nmds <- as.data.frame(veg_nmds)


zeigerwerte_pa <- veg_nmds %>%
  dplyr::mutate(across(where(is.numeric), ~ ifelse(. > 0, 1, 0)))

zeigerwerte_pa <- t(zeigerwerte_pa)
zeigerwerte_pa <- as.data.frame(zeigerwerte_pa)
zeigerwerte_pa <- zeigerwerte_pa %>% tibble::rownames_to_column("species")
zeigerwerte_pa <- zeigerwerte_pa %>%
  mutate(across(-species, as.numeric))  # Alle Spalten außer "species" numerisch machen
unique(zeigerwerte_pa$species)  # Gibt alle eindeutigen Werte in der Spalte zurück
```


## Lichtzahl
```{r}
lichtzahlen$Lichtzahl <- as.numeric(lichtzahlen$Lichtzahl)

zeigerwerte_licht <- zeigerwerte_pa %>%
  pivot_longer(
    cols = -species,  # Alle Spalten außer "species"
    names_to = "plot",
    values_to = "presence"
  ) %>%
  left_join(lichtzahlen, by = "species") %>%  # Lichtzahlen hinzufügen
  mutate(value = ifelse(presence == 1, Lichtzahl, 0)) %>%  # Werte setzen
  dplyr::select(-presence, -Lichtzahl) %>%  # Unnötige Spalten entfernen
  pivot_wider(names_from = plot, values_from = value)


#zeigerwerte_licht[] <- lapply(zeigerwerte_licht, as.numeric)
#Alle 0 auf NA setzen, damit diese nicht für die Mittelwerte mitberechnet werden
zeigerwerte_licht[zeigerwerte_licht == 0] <- NA
zeigerwerte_licht <- as.data.frame(zeigerwerte_licht)
zeigerwerte_licht <- zeigerwerte_licht %>% column_to_rownames("species")
zeigerwerte_licht <- t(zeigerwerte_licht)
zeigerwerte_licht <- as.data.frame(zeigerwerte_licht)

zeigerwerte_licht <- zeigerwerte_licht %>%
  rownames_to_column("plot") %>%                # Zeilennamen in Spalte verschieben
  mutate(across(-plot, ~ na_if(., 0))) %>%      # Nur numerische Spalten bearbeiten
  column_to_rownames("plot")                    # Spalte wieder in Zeilennamen umwandeln


zeigerwerte_licht_mean <- zeigerwerte_licht %>%
  rownames_to_column("plot") %>%                      # Zeilennamen in eine Spalte verschieben
  rowwise() %>%                                       # Zeilenweise Berechnung aktivieren
  mutate(mean_value = mean(c_across(-plot), na.rm = TRUE)) %>%  # Mittelwert berechnen
  ungroup() %>%                                       # Gruppierung aufheben
  column_to_rownames("plot")                          # Spalte wieder in Zeilennamen umwandeln

#Rownames als Spalte für Excel
zeigerwerte_licht_mean_ex = zeigerwerte_licht_mean %>%
                 rownames_to_column("plot")

# Alphabetisch sortieren
zeigerwerte_licht_mean_ex <- zeigerwerte_licht_mean_ex[order(zeigerwerte_licht_mean_ex$plot), ] 
```



## Temperaturzahl
```{r}
temperaturzahlen$Temperaturzahl <- as.numeric(temperaturzahlen$Temperaturzahl)


zeigerwerte_temp <- zeigerwerte_pa %>%
  pivot_longer(
    cols = -species,  # Alle Spalten außer "species"
    names_to = "plot",
    values_to = "presence"
  ) %>%
  left_join(temperaturzahlen, by = "species") %>%  # Temperaturzahl hinzufügen
  mutate(value = ifelse(presence == 1, Temperaturzahl, 0)) %>%  # Werte setzen
  dplyr::select(-presence, -Temperaturzahl) %>%  # Unnötige Spalten entfernen
  pivot_wider(names_from = plot, values_from = value)


#zeigerwerte_temp[] <- lapply(zeigerwerte_temp, as.numeric)
#Alle 0 auf NA setzen, damit diese nicht für die Mittelwerte mitberechnet werden
zeigerwerte_temp[zeigerwerte_temp == 0] <- NA
zeigerwerte_temp <- as.data.frame(zeigerwerte_temp)
zeigerwerte_temp <- zeigerwerte_temp %>% column_to_rownames("species")
zeigerwerte_temp <- t(zeigerwerte_temp)
zeigerwerte_temp <- as.data.frame(zeigerwerte_temp)



zeigerwerte_temp <- zeigerwerte_temp %>%
  rownames_to_column("plot") %>%                # Zeilennamen in Spalte verschieben
  mutate(across(-plot, ~ na_if(., 0))) %>%      # Nur numerische Spalten bearbeiten
  column_to_rownames("plot")                    # Spalte wieder in Zeilennamen umwandeln

zeigerwerte_temp_mean <- zeigerwerte_temp %>%
  rownames_to_column("plot") %>%                      # Zeilennamen in eine Spalte verschieben
  rowwise() %>%                                       # Zeilenweise Berechnung aktivieren
  mutate(mean_value = mean(c_across(-plot), na.rm = TRUE)) %>%  # Mittelwert berechnen
  ungroup() %>%                                       # Gruppierung aufheben
  column_to_rownames("plot")                                     # Gruppierung aufheben

#Rownames als Spalte für Excel
zeigerwerte_temp_mean_ex <- zeigerwerte_temp_mean %>%
  rownames_to_column("plot")
# Alphabetisch sortieren
zeigerwerte_temp_mean_ex <- zeigerwerte_temp_mean_ex[order(zeigerwerte_temp_mean_ex$plot), ] 
```


## Kontinentalitaetszahl
```{r}
kontizahlen$Kontinentalitaetszahl <- as.numeric(kontizahlen$Kontinentalitaetszahl)


zeigerwerte_konti <- zeigerwerte_pa %>%
  pivot_longer(
    cols = -species,  # Alle Spalten außer "species"
    names_to = "plot",
    values_to = "presence"
  ) %>%
  left_join(kontizahlen, by = "species") %>%  # Kontinentalitaetszahl hinzufügen
  mutate(value = ifelse(presence == 1, Kontinentalitaetszahl, 0)) %>%  # Werte setzen
  dplyr::select(-presence, -Kontinentalitaetszahl) %>%  # Unnötige Spalten entfernen
  pivot_wider(names_from = plot, values_from = value)

#zeigerwerte_temp[] <- lapply(zeigerwerte_temp, as.numeric)
#Alle 0 auf NA setzen, damit diese nicht für die Mittelwerte mitberechnet werden
zeigerwerte_konti[zeigerwerte_konti == 0] <- NA
zeigerwerte_konti <- as.data.frame(zeigerwerte_konti)
zeigerwerte_konti <- zeigerwerte_konti %>% column_to_rownames("species")
zeigerwerte_konti <- t(zeigerwerte_konti)
zeigerwerte_konti <- as.data.frame(zeigerwerte_konti)


zeigerwerte_konti <- zeigerwerte_konti %>%
  rownames_to_column("plot") %>%                # Zeilennamen in Spalte verschieben
  mutate(across(-plot, ~ na_if(., 0))) %>%      # Nur numerische Spalten bearbeiten
  column_to_rownames("plot")                    # Spalte wieder in Zeilennamen umwandeln

zeigerwerte_konti_mean <- zeigerwerte_konti %>%
  rownames_to_column("plot") %>%                      # Zeilennamen in eine Spalte verschieben
  rowwise() %>%                                       # Zeilenweise Berechnung aktivieren
  mutate(mean_value = mean(c_across(-plot), na.rm = TRUE)) %>%  # Mittelwert berechnen
  ungroup() %>%                                       # Gruppierung aufheben
  column_to_rownames("plot")        

#Rownames als Spalte für Excel
zeigerwerte_konti_mean_ex <- zeigerwerte_konti_mean %>%
  rownames_to_column("plot")
zeigerwerte_konti_mean_ex <- zeigerwerte_konti_mean_ex[order(zeigerwerte_konti_mean_ex$plot), ] # Alphabetisch sortieren
```



## Feuchtezahl
```{r}
feuchtezahlen$Feuchtezahl <- as.numeric(feuchtezahlen$Feuchtezahl)

zeigerwerte_feucht <- zeigerwerte_pa %>%
  pivot_longer(
    cols = -species,  # Alle Spalten außer "species"
    names_to = "plot",
    values_to = "presence"
  ) %>%
  left_join(feuchtezahlen, by = "species") %>%  # Feuchtezahl hinzufügen
  mutate(value = ifelse(presence == 1, Feuchtezahl, 0)) %>%  # Werte setzen
  dplyr::select(-presence, -Feuchtezahl) %>%  # Unnötige Spalten entfernen
  pivot_wider(names_from = plot, values_from = value)

#zeigerwerte_temp[] <- lapply(zeigerwerte_temp, as.numeric)
#Alle 0 auf NA setzen, damit diese nicht für die Mittelwerte mitberechnet werden
zeigerwerte_feucht[zeigerwerte_feucht == 0] <- NA
zeigerwerte_feucht <- as.data.frame(zeigerwerte_feucht)
zeigerwerte_feucht <- zeigerwerte_feucht %>% column_to_rownames("species")
zeigerwerte_feucht <- t(zeigerwerte_feucht)
zeigerwerte_feucht <- as.data.frame(zeigerwerte_feucht)

zeigerwerte_feucht <- zeigerwerte_feucht %>%
  rownames_to_column("plot") %>%                # Zeilennamen in Spalte verschieben
  mutate(across(-plot, ~ na_if(., 0))) %>%      # Nur numerische Spalten bearbeiten
  column_to_rownames("plot")                    # Spalte wieder in Zeilennamen umwandeln

zeigerwerte_feucht_mean <- zeigerwerte_feucht %>%
  rownames_to_column("plot") %>%                      # Zeilennamen in eine Spalte verschieben
  rowwise() %>%                                       # Zeilenweise Berechnung aktivieren
  mutate(mean_value = mean(c_across(-plot), na.rm = TRUE)) %>%  # Mittelwert berechnen
  ungroup() %>%                                       # Gruppierung aufheben
  column_to_rownames("plot")        

#Rownames als Spalte für Excel
zeigerwerte_feucht_mean_ex <- zeigerwerte_feucht_mean %>%
  rownames_to_column("plot")

# Alphabetisch sortieren
zeigerwerte_feucht_mean_ex <- zeigerwerte_feucht_mean_ex[order(zeigerwerte_feucht_mean_ex$plot), ] 
```



## Reaktionszahl
```{r}
reaktionszahlen$Reaktionszahl <- as.numeric(reaktionszahlen$Reaktionszahl)

zeigerwerte_reakt <- zeigerwerte_pa %>%
  pivot_longer(
    cols = -species,  # Alle Spalten außer "species"
    names_to = "plot",
    values_to = "presence"
  ) %>%
  left_join(reaktionszahlen, by = "species") %>%  # Reaktionszahl hinzufügen
  mutate(value = ifelse(presence == 1, Reaktionszahl, 0)) %>%  # Werte setzen
  dplyr::select(-presence, -Reaktionszahl) %>%  # Unnötige Spalten entfernen
  pivot_wider(names_from = plot, values_from = value)

#zeigerwerte_temp[] <- lapply(zeigerwerte_temp, as.numeric)
#Alle 0 auf NA setzen, damit diese nicht für die Mittelwerte mitberechnet werden
zeigerwerte_reakt[zeigerwerte_reakt == 0] <- NA
zeigerwerte_reakt <- as.data.frame(zeigerwerte_reakt)
zeigerwerte_reakt <- zeigerwerte_reakt %>% column_to_rownames("species")
zeigerwerte_reakt <- t(zeigerwerte_reakt)
zeigerwerte_reakt <- as.data.frame(zeigerwerte_reakt)


zeigerwerte_reakt <- zeigerwerte_reakt %>%
  rownames_to_column("plot") %>%                # Zeilennamen in Spalte verschieben
  mutate(across(-plot, ~ na_if(., 0))) %>%      # Nur numerische Spalten bearbeiten
  column_to_rownames("plot")                    # Spalte wieder in Zeilennamen umwandeln

zeigerwerte_reakt_mean <- zeigerwerte_reakt %>%
  rownames_to_column("plot") %>%                      # Zeilennamen in eine Spalte verschieben
  rowwise() %>%                                       # Zeilenweise Berechnung aktivieren
  mutate(mean_value = mean(c_across(-plot), na.rm = TRUE)) %>%  # Mittelwert berechnen
  ungroup() %>%                                       # Gruppierung aufheben
  column_to_rownames("plot") 

#Rownames als Spalte für Excel
zeigerwerte_reakt_mean_ex <- zeigerwerte_reakt_mean %>%
  rownames_to_column("plot")

# Alphabetisch sortieren
zeigerwerte_reakt_mean_ex <- zeigerwerte_reakt_mean_ex[order(zeigerwerte_reakt_mean_ex$plot), ] 
```



## Naehrstoffzahl
```{r}
naehrstoffzahlen$Naehrstoffzahl <- as.numeric(naehrstoffzahlen$Naehrstoffzahl)

zeigerwerte_naehr <- zeigerwerte_pa %>%
  pivot_longer(
    cols = -species,  # Alle Spalten außer "species"
    names_to = "plot",
    values_to = "presence"
  ) %>%
  left_join(naehrstoffzahlen, by = "species") %>%  # Reaktionszahl hinzufügen
  mutate(value = ifelse(presence == 1, Naehrstoffzahl, 0)) %>%  # Werte setzen
  dplyr::select(-presence, -Naehrstoffzahl) %>%  # Unnötige Spalten entfernen
  pivot_wider(names_from = plot, values_from = value)

#zeigerwerte_temp[] <- lapply(zeigerwerte_temp, as.numeric)
#Alle 0 auf NA setzen, damit diese nicht für die Mittelwerte mitberechnet werden
zeigerwerte_naehr[zeigerwerte_naehr == 0] <- NA
zeigerwerte_naehr <- as.data.frame(zeigerwerte_naehr)
zeigerwerte_naehr <- zeigerwerte_naehr %>% column_to_rownames("species")
zeigerwerte_naehr <- t(zeigerwerte_naehr)
zeigerwerte_naehr <- as.data.frame(zeigerwerte_naehr)


zeigerwerte_naehr <- zeigerwerte_naehr %>%
  rownames_to_column("plot") %>%                # Zeilennamen in Spalte verschieben
  mutate(across(-plot, ~ na_if(., 0))) %>%      # Nur numerische Spalten bearbeiten
  column_to_rownames("plot")                    # Spalte wieder in Zeilennamen umwandeln

zeigerwerte_naehr_mean <- zeigerwerte_naehr %>%
  rownames_to_column("plot") %>%                      # Zeilennamen in eine Spalte verschieben
  rowwise() %>%                                       # Zeilenweise Berechnung aktivieren
  mutate(mean_value = mean(c_across(-plot), na.rm = TRUE)) %>%  # Mittelwert berechnen
  ungroup() %>%                                       # Gruppierung aufheben
  column_to_rownames("plot") 


#Rownames als Spalte für Excel
zeigerwerte_naehr_mean_ex <- zeigerwerte_naehr_mean %>%
  rownames_to_column("plot")

# Alphabetisch sortieren
zeigerwerte_naehr_mean_ex <- zeigerwerte_naehr_mean_ex[order(zeigerwerte_naehr_mean_ex$plot), ] 
```


## Salzzahl
```{r}
salzzahlen$Salzzahl <- as.numeric(salzzahlen$Salzzahl)

zeigerwerte_salz <- zeigerwerte_pa %>%
  pivot_longer(
    cols = -species,  # Alle Spalten außer "species"
    names_to = "plot",
    values_to = "presence"
  ) %>%
  left_join(salzzahlen, by = "species") %>%  # Reaktionszahl hinzufügen
  mutate(value = ifelse(presence == 1, Salzzahl, 0)) %>%  # Werte setzen
  dplyr::select(-presence, -Salzzahl) %>%  # Unnötige Spalten entfernen
  pivot_wider(names_from = plot, values_from = value)

#zeigerwerte_temp[] <- lapply(zeigerwerte_temp, as.numeric)
#Alle 0 auf NA setzen, damit diese nicht für die Mittelwerte mitberechnet werden
zeigerwerte_salz[zeigerwerte_salz == 0] <- NA
zeigerwerte_salz <- as.data.frame(zeigerwerte_salz)
zeigerwerte_salz <- zeigerwerte_salz %>% column_to_rownames("species")
zeigerwerte_salz <- t(zeigerwerte_salz)
zeigerwerte_salz <- as.data.frame(zeigerwerte_salz)


zeigerwerte_salz <- zeigerwerte_salz %>%
  rownames_to_column("plot") %>%                # Zeilennamen in Spalte verschieben
  mutate(across(-plot, ~ na_if(., 0))) %>%      # Nur numerische Spalten bearbeiten
  column_to_rownames("plot")                    # Spalte wieder in Zeilennamen umwandeln

zeigerwerte_salz_mean <- zeigerwerte_salz %>%
  rownames_to_column("plot") %>%                      # Zeilennamen in eine Spalte verschieben
  rowwise() %>%                                       # Zeilenweise Berechnung aktivieren
  mutate(mean_value = mean(c_across(-plot), na.rm = TRUE)) %>%  # Mittelwert berechnen
  ungroup() %>%                                       # Gruppierung aufheben
  column_to_rownames("plot")

#Rownames als Spalte für Excel
zeigerwerte_salz_mean_ex <- zeigerwerte_salz_mean %>%
  rownames_to_column("plot")

# Alphabetisch sortieren
zeigerwerte_salz_mean_ex <- zeigerwerte_salz_mean_ex[order(zeigerwerte_salz_mean_ex$plot), ] 
```






# RDA preparation
## RDA data
```{r}
env_RDA <- plot_data %>% 
  select(Plot_ID, Elevation, Cover_Juvenile) %>% 
  column_to_rownames(var = "Plot_ID")

covar_RDA <- plot_data %>% 
  select(Plot_ID, Alm_ID) %>% 
  column_to_rownames(var = "Plot_ID")

# install.packages("ade4")
library(ade4)
covar_RDA <- acm.disjonctif(covar_RDA) #den Faktor Alm als Dummy-Variable (siehe auch https://stackoverflow.com/questions/5048638/automatically-expanding-an-r-factor-into-a-collection-of-1-0-indicator-variables)

spec_RDA.sqrt <- sqrt(veg_layer_red)

```

## Data to relate scores with

### Combine vegetation data with bioflor
```{r}
df <- subset(veg, layer %in% c("hl", "sl"))
df <- data.frame(df, row.names = NULL)
rownames(df) <- df[,1]
df <- df[, -c(1, 2)]

df1<- df %>%
  mutate(species = rownames(.)) %>%      # 1. Add rownames as a column
  pivot_longer(
    cols = -species,                     # 2. Keep 'species', pivot the rest
    names_to = "plot_id",                # 3. Column names become 'plot_id'
    values_to = "cover"                  # 4. Values go into 'cover'
  )

df2 = bioflorBiotop

traits = merge(df1,df2,by="species",all.x=TRUE)
```

### Forest association and biotope
```{r}
df <- traits %>%
  mutate(forestAssociation = fct_explicit_na(forestAssociation, na_level = "Others"),  # ensure NA is labeled
         forestAssociation = ifelse(forestAssociation == "Others",
                          "Others",
                          substr(as.character(forestAssociation), 
                                 2, nchar(as.character(forestAssociation)))))  # remove first letter

# Reorder factor levels
df$forestAssociation <- factor(df$forestAssociation, 
                               levels = c("1.1", "1.2", "2.1","2.2","Others"))


# combine tables 34,35 if presence then 1, otherwise 0
df = df %>% dplyr::mutate(
   tables34_35 = case_when(
      !is.na(table34) & !is.na(table35) ~ 1,  # both present → 1
      !is.na(table34) ~ 1,
      !is.na(table35) ~ 1,
      TRUE ~ NA_real_    # neither present
    ))

df$tables34_35[is.na(df$tables34_35)] <- 0

df = df %>% dplyr::mutate(
   Others = case_when(
      is.na(table33) &  is.na(table34) & is.na(table35) ~ 1, 
      TRUE ~ NA_real_    # neither present
    ))

df$Others[is.na(df$Others)] <- 0

df$table33 = ifelse(df$table33 > 0, 1, 0)

df$table33[is.na(df$table33)] <- 0



# filter data for only species presences
df %<>% dplyr::filter(cover>0)


traitsBiotopeForest = df
```


### weight values
```{r}
traits_weighted <- traitsBiotopeForest %>% 
  dplyr::mutate(mowingTolerance_weighted = cover*mowingTolerance, # jf NA then NA
         grazingTolerance_weighted = cover*grazingTolerance, # jf NA then NA
         tramplingTolerance_weighted = cover*tramplingTolerance, # jf NA then NA
         forageValue_weighted = cover*forageValue) %>% # jf NA then NA
  dplyr::select(plot_id,species, mowingTolerance_weighted,
                grazingTolerance_weighted,tramplingTolerance_weighted,
                forageValue_weighted)

traitsSummaryPlot <- traits_weighted %>% 
  dplyr::group_by(plot_id) %>% 
  dplyr::summarise(richness = n_distinct(species),
                   mowingTolerance_weighted = 
                     sum(mowingTolerance_weighted, na.rm = T)/100,
                   grazingTolerance_weighted = 
                     sum(grazingTolerance_weighted, na.rm = T)/100,
                   tramplingTolerance_weighted = 
                     sum(tramplingTolerance_weighted, na.rm = T)/100,
                   forageValue_weighted = 
                     sum(forageValue_weighted, na.rm = T)/100)
```



# save
```{r}
save(
  plot_data,coords,
  veg, GermanSl,veg_layer_red,
  treatment,
  veg_layer,veg_layer_red,vegTreatment,
  
  zeigerwerte,zeigerwerte_mittel,
  
  zeigerwerte_salz_mean_ex,zeigerwerte_naehr_mean_ex,
  zeigerwerte_reakt_mean_ex,zeigerwerte_konti_mean_ex,
  zeigerwerte_licht_mean_ex,zeigerwerte_temp_mean_ex,
  zeigerwerte_feucht_mean_ex,
  

 # RDA
 env_RDA, covar_RDA,spec_RDA.sqrt,
 waldarten,rl_bayern,bioflorBiotop,traitsSummaryPlot,traits,traitsBiotopeForest,
  file="data/workingdata.RData")
```
