==> r-code/05_1_ordinations.Rmd <==
---
title: "Ordination analyses"
author: "Bernd Panassiti"
date: 'created: 28.11.2024, last modified: `r format(Sys.Date(), format="%d.%m.%Y")`'
output: pdf_document
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
library(knitr)
opts_knit$set(root.dir='../')                     # definining working directory; or normalizePath('../')
opts_chunk$set(fig.align='center',                # aligns all figures
                echo=TRUE,                        # shows r-code
                message=FALSE,                    # suppresses library outputs
                warnings=FALSE,                   # suppresses library outputs
                tidy=TRUE,  # prevents the source code from running off a pdf page
                dev='pdf')
```


# Load packages
```{r}
#Install/load pacman
if(!require(pacman)){install.packages("pacman");require(pacman)}
#Install packages
p_load(tidyverse,dplyr,vegan,ggplot2,magrittr,multcompView)
```

# load data and functions
```{r}
load("data/workingdata.RData")

# Scale observations by species cover
source("r-code/00_functions.r")
```


# NMDS
## Run NMDS
```{r}
rm(.Random.seed, envir=globalenv())
set.seed(2025)

nmds <-  vegan::metaMDS(veg_layer_red)
```

## Umweltparameter hinzufügen
```{r}
df = plot_data
plot_data_sorted <- df[match(rownames(veg_layer_red), plot_data$Plot_ID), ]


fit <- envfit(nmds, plot_data_sorted[, c("Elevation", "Cover_Juvenile", "Hangneigung")], permutations = 999)

plot(nmds, type = "n")  # Leerer Plot der NMDS
points(nmds, display = "sites", pch = 21, bg = "lightblue", cex = 1.5)  # Punkte hinzufügen
plot(fit, col = "red", lwd = 2) 

nmds_points <- as.data.frame(scores(nmds, display = "sites"))
nmds_points$Site <- rownames(nmds_points)
nmds_points <- nmds_points[order(rownames(nmds_points)), ]

nmds_points$Verbuschung <- plot_data$Verbuschung[match(nmds_points$Site, plot_data$Plot_ID)]
nmds_points$Verbuschung <- ifelse(nmds_points$Verbuschung == 0, "unverbuscht", "verbuscht")
nmds_points$Elevation <- plot_data$Elevation[match(nmds_points$Site, plot_data$Plot_ID)]
nmds_points$Elevation <- as.numeric(plot_data$Elevation)
# Vektor aus envfit extrahieren
vector_fit <- as.data.frame(fit$vectors$arrows)
vector_fit$Variable <- rownames(vector_fit)

# ggplot2-Plot
vector_fit$label_x <- vector_fit$NMDS1 + 0.19
```


## NMDS mit Verbuschung und Umweltparametern
### Plot nur mit Verbuschung
```{r}
ggplot(nmds_points, aes(x = NMDS1, y = NMDS2, color = Verbuschung)) +
  geom_point(size = 3) +
  scale_size_continuous(range = c(3, 10)) +
  scale_color_manual(values = c("verbuscht" = "red", "unverbuscht" = "darkgreen")) +
  theme_minimal() +
  labs(title = "NMDS mit Verbuschung", x = "NMDS1", y = "NMDS2")
```


### Mit Plotbenennung
```{r}
ggplot(nmds_points, aes(x = NMDS1, y = NMDS2, color = Verbuschung)) +
  geom_point(size = 3) +
  scale_size_continuous(range = c(3, 10)) +
  scale_color_manual(values = c("verbuscht" = "red", "unverbuscht" = "darkgreen")) +
  geom_segment(data = vector_fit, aes(x = 0, y = 0, xend = NMDS1, yend = NMDS2),
               arrow = arrow(length = unit(0.2, "cm")), color = "black", linewidth = 1) +
  geom_text(data = vector_fit, aes(x = label_x, y = NMDS2, label = Variable),
            color = "black", vjust = 0.1) +
   geom_text(data=nmds_points,aes(x=NMDS1,y=NMDS2,label=Site), vjust=-1) +
  theme_minimal() +
  labs(title = "NMDS mit Umweltparametern", x = "NMDS1", y = "NMDS2")
```


### Ohne Plotbenennung
```{r}
ggplot(nmds_points, aes(x = NMDS1, y = NMDS2, color = Verbuschung)) +
  geom_point(size = 3) +
  scale_size_continuous(range = c(3, 10)) +
  scale_color_manual(values = c("verbuscht" = "red", "unverbuscht" = "darkgreen")) +
  geom_segment(data = vector_fit, aes(x = 0, y = 0, xend = NMDS1, yend = NMDS2),
               arrow = arrow(length = unit(0.2, "cm")), color = "black", size = 1) +
  geom_text(data = vector_fit, aes(x = label_x, y = NMDS2, label = Variable),
            color = "black", vjust = 0.1) +
  theme_minimal() +
  labs(title = "NMDS mit Umweltparametern", x = "NMDS1", y = "NMDS2")
```

### Mit Skalierung der Punktgröße nach Höhe
#### Mit Plotbenennung
```{r}
ggplot(nmds_points, aes(x = NMDS1, y = NMDS2, color = Verbuschung, size = Elevation)) +
  geom_point() +
  scale_size_continuous(range = c(3, 10)) +
  scale_color_manual(values = c("verbuscht" = "red", "unverbuscht" = "darkgreen")) +
  geom_segment(data = vector_fit, aes(x = 0, y = 0, xend = NMDS1, yend = NMDS2),
               arrow = arrow(length = unit(0.2, "cm")), color = "black", size = 1) +
  geom_text(data=nmds_points,aes(x=NMDS1,y=NMDS2,label=Site), vjust=-1) +
  theme_minimal() +
  labs(title = "NMDS mit Umweltparametern", x = "NMDS1", y = "NMDS2")
```


#### Ohne Plotbenennung
```{r NMDS mit Verbuschung}

ggplot(nmds_points, aes(x = NMDS1, y = NMDS2, color = Verbuschung, size = Elevation)) +
  geom_point() +
  scale_size_continuous(range = c(3, 10)) +
  scale_color_manual(values = c("verbuscht" = "red", "unverbuscht" = "darkgreen")) +
  geom_segment(data = vector_fit, aes(x = 0, y = 0, xend = NMDS1, yend = NMDS2),
               arrow = arrow(length = unit(0.2, "cm")), color = "black", size = 1) +
  theme_minimal() +
  labs(title = "NMDS mit Umweltparametern", x = "NMDS1", y = "NMDS2")
```




# NMDS mit Zeigerwerten


fit <- envfit(nmds, zeigerwerte_mittel[, c("Lichtzahl", "Temperaturzahl", "Kontinentalitaetszahl", "Feuchtezahl", "Reaktionszahl", "Naehrstoffzahl")], permutations = 999)

Hier werden nur Herblayer und Seedling Layer verwendet.
## Data preparation
```{r}
df1 = veg_layer_red
df2 = plot_data

df1 <- df1[order(rownames(df1)), ]
df2 <- df2[order(df2$Plot_ID), ]

df2 %<>% dplyr::rename(treatment = Verbuschung,
                       encroachment = Cover_Juvenile)


# ensure that row order is the same in both data.frames
df2_sorted <- df2[match(rownames(df1), df2$Plot_ID), ]


df3 = zeigerwerte_mittel %>% 
  dplyr::rename(light = Lichtzahl, 
                temperature = Temperaturzahl, 
                continentality = Kontinentalitaetszahl, 
                moisture = Feuchtezahl, 
                reaction = Reaktionszahl, 
                nutrient = Naehrstoffzahl)

#Cover_Juvenile zu zeigerwerte_mittel hinzufügen
indicatorValues_mean <- merge(
  df3, 
  df2_sorted[, c("Plot_ID", "treatment", "encroachment")], 
  by = "Plot_ID", 
  all.x = TRUE
)
```



## Run NMDS
```{r NMDS mit Zeigerwerten}
rm(.Random.seed, envir=globalenv())
set.seed(2025)

nmds <- vegan::metaMDS(df1)


fit <- envfit(nmds, indicatorValues_mean[, c("light",
                                             "temperature",
                                             "continentality",
                                             "moisture",
                                             "reaction",
                                             "nutrient")], permutations = 999)


nmds_points <- as.data.frame(scores(nmds, display = "sites"))
nmds_points$Site <- rownames(nmds_points)
veg_Verbuschung <- data.frame(rownames(veg_layer), df2_sorted[, 10])
colnames(veg_Verbuschung) = c("Site", "treatment")

indicatorValues_mean_nmds <- indicatorValues_mean %>%
  rename(Site = Plot_ID) 

nmds_points_merge <- merge(
  nmds_points, 
  indicatorValues_mean_nmds[, c("Site", "encroachment","treatment")], 
  by = "Site", all.x = TRUE)

nmds_points_merge$treatment <- ifelse(nmds_points_merge$treatment == 1, "encroachment", "control")
nmds_points_merge$treatment <- as.factor(nmds_points_merge$treatment) 

# Vektor aus envfit extrahieren mit Kontinentalitätszahl
vector_fit <- as.data.frame(fit$vectors$arrows)
vector_fit$Variable <- rownames(vector_fit)

vector_fit$label_x <- vector_fit$NMDS1 + 0.19
```

## Summary
```{r}
# Basic summary
nmds

# Stress value
stress_value=my_signif(nmds$stress);stress_value

# Scores (site and species)
scores(nmds)

# Environmental fit (optional, if you have env variables)
# Assume env_data contains e.g., elevation, slope, etc.
fit
```



## Plot NMDS
```{r}
Appendix_S6 = ggplot() +
  geom_point(
    data = nmds_points_merge,
    aes(x = NMDS1, y = NMDS2, 
        color = treatment, 
        size = encroachment) ) +
  scale_color_manual("Treatments:",
        values =  c("#d9d9d9", "#969696"))+
  scale_size_continuous("Encroachment gradient (%):",
                        range = c(3, 10)) +
  geom_segment(data = vector_fit, 
               aes(x = 0, y = 0, 
                   xend = NMDS1, 
                   yend = NMDS2),
               arrow = arrow(
                 length = unit(0.2, "cm")), 
               color = "black", size = 1) +
   geom_text(data = vector_fit, 
             aes(x = NMDS1,
                 y = NMDS2, 
                 label = Variable),
            color = "black", 
            vjust = 0.1) +
  theme_minimal() +
  labs(title ="", 
       x = "NMDS1", 
       y = "NMDS2")+
  annotate("text", x = Inf, y = -Inf,
           label = paste("Stress =", stress_value),
           hjust = 1.1, vjust = -0.5, size = 4);Appendix_S6
```

# save
```{r}
save(
  Appendix_S6,
  file="results/ordinations_NMDS.RData"
)
```
