==> r-code/05_2_ordinations-RDA.Rmd <==
---
title: "RDA and Variation partitioning"
author: "Bernd Panassiti"
output: html_document
date: 'created: 23.04.2025, last modified: `r format(Sys.Date(), format="%d.%m.%Y")`'
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r setup, include=FALSE}
library(knitr)
opts_knit$set(root.dir='../')                     # definining working directory; or normalizePath('../')
opts_chunk$set(fig.align='center',                # aligns all figures
                echo=TRUE,                        # shows r-code
                message=FALSE,                    # suppresses library outputs
                warnings=FALSE,                   # suppresses library outputs
                tidy=TRUE,  # prevents the source code from running off a pdf page
                dev='pdf')
```

# load packages
```{r}
#Install/load pacman
if(!require(pacman)){install.packages("pacman");require(pacman)}
#Install packages
p_load(tidyverse,dplyr,vegan,ggplot2,ggpubr,magrittr,multcompView,FSA,
       forcats # für fct_recode()
       )
```

# load working data
```{r}
load("data/workingdata.RData")
```


# RDA

RDA: simpler methods than CCA, variation partitioning (all other variables included as cofactors and influence subtracted)

How much variance in species composition can be attributed to scrub encroachment

Winners and losers of scrub encroachment
Evaluate forest species
Pasture species/ grassland


```{r}
rm(.Random.seed, envir=globalenv())
set.seed(2025)

RDA <- rda(spec_RDA.sqrt, env_RDA[,-1], covar_RDA)
summary(RDA)

anova.cca(RDA, step = 1000) #globaler Test des RDA Ergebnisses
#-> das Ergebnis der RDA ist signifikant

ordiplot(RDA) #Ordiplot

smry <- summary(RDA)
RDAscoresSites  <- data.frame(smry$sites[,1:2])       # sites
RDAscoresSpecies  <- data.frame(smry$species[,1:2])     # species

```

1. RDA axis related to woody encroachment

# Variation Partioning
```{r}
rm(.Random.seed, envir=globalenv())
set.seed(2025)

env_RDA.CovJuv <- env_RDA %>% 
  dplyr::select(Cover_Juvenile)

spe.part.all <- vegan::varpart(spec_RDA.sqrt, 
                        env_RDA.CovJuv,
                        covar_RDA)

plot(spe.part.all,
     Xnames = c( "Cover Juvenile", "Alm"), # name the partitions
     bg = c("seagreen3",
            "mediumpurple"), alpha = 80, # colour the circles
     digits = 2, # only show 2 digits
     cex = 1.5) # Venn diagramm mit Farben und Beschriftung

print(spe.part.all)
```



Partition of variance in RDA 

Call: vegan::varpart(Y = spec_RDA.sqrt, X = env_RDA.CovJuv, covar_RDA)

Explanatory tables:
X1:  env_RDA.CovJuv
X2:  covar_RDA 

No. of explanatory tables: 2 
Total variation (SS): 9110.5 
            Variance: 63.71 
No. of observations: 144 

Partition table:
                     Df R.squared Adj.R.squared Testable
[a+c] = X1            1   0.02490       0.01803     TRUE
[b+c] = X2            7   0.37801       0.34600     TRUE
[a+b+c] = X1+X2       8   0.40185       0.36640     TRUE
Individual fractions                                    
[a] = X1|X2           1                 0.02040     TRUE
[b] = X2|X1           7                 0.34837     TRUE
[c]                   0                -0.00237    FALSE
[d] = Residuals                         0.63360    FALSE
---
Use function ‘rda’ to test significance of fractions of interest

```{r}
rm(.Random.seed, envir=globalenv())
set.seed(2025)

# For environmental variables
rda_env <- rda(spec_RDA.sqrt ~ ., data = env_RDA.CovJuv)
anova(rda_env)

# For covariables
rda_cov <- rda(spec_RDA.sqrt ~ ., data = covar_RDA)
anova(rda_cov)
```


> rda_env <- rda(spec_RDA.sqrt ~ ., data = env_RDA.CovJuv)
> anova(rda_env)
Permutation test for rda under reduced model
Permutation: free
Number of permutations: 999

Model: rda(formula = spec_RDA.sqrt ~ Cover_Juvenile, data = env_RDA.CovJuv)
          Df Variance      F Pr(>F)    
Model      1    1.586 3.6261  0.001 ***
Residual 142   62.124                  
---
Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
> 
> # For covariables
> rda_cov <- rda(spec_RDA.sqrt ~ ., data = covar_RDA)

Some constraints or conditions were aliased because they were redundant. This can happen if terms are linearly dependent (collinear):
‘Alm_ID.s’
> anova(rda_cov)
Permutation test for rda under reduced model
Permutation: free
Number of permutations: 999

Model: rda(formula = spec_RDA.sqrt ~ Alm_ID.a + Alm_ID.b + Alm_ID.e + Alm_ID.g + Alm_ID.h + Alm_ID.k + Alm_ID.m + Alm_ID.s, data = covar_RDA)
          Df Variance      F Pr(>F)    
Model      7   24.083 11.808  0.001 ***
Residual 136   39.627                  
---
Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

# Dataframe - Agriculture parameter


```{r}
df = RDAscoresSites
df$plot_id = rownames(RDAscoresSites)

RDAscoresSitesTraits <- merge(df, traitsSummaryPlot, by = "plot_id", all.x = TRUE)
```




# Forage value
Linear Regression: RDA1 vs Forage Value

```{r}
df = RDAscoresSitesTraits %>%
  mutate(forageValue = as.numeric(forageValue_weighted)) %>%      # convert FW to numeric
  rename(`Forage Value` = forageValue,
         `RDA1-Value` = RDA1) %>%      # rename for plotting
  filter(!is.na(`Forage Value`), !is.na(`RDA1-Value`))  # remove NA values


gRDA_forage = ggplot(df, aes(y = `Forage Value`, x = `RDA1-Value`)) +
  geom_point(color = "steelblue") +
  geom_smooth(method = "lm", se = TRUE, color = "darkred", fill = "pink") +
  stat_regline_equation(
    aes(label = after_stat(eq.label)),
    label.x = -3,
    label.y = max(df$`Forage Value`, na.rm = TRUE)
  ) +
stat_cor(
  aes(label = after_stat(paste(rr.label, p.label, sep = "~`,`~"))),
  label.x = -3,
  label.y = max(df$`Forage Value`, na.rm = TRUE) - 0.2,
  p.accuracy = 0.001
)+
  theme_minimal() +
  labs(
    title = "",
    y = "Forage value (weighted)",
    x = "RDA1 - Site score (Reverse of encroachment)"
  ); gRDA_forage
```





# Forest association

Waldartenliste
1.1: geschlossener Wald
1.2: Waldränder und –verlichtungen
2.1: Wald wie im Offenland
2.2: auch Wald, aber Schwerpunkt Offenland

B - Bäume, S - Sträucher, K - Kräuter, E - Epiphyten

### Data preparation

```{r}
# Schritt 1: Merge Daten
df1 <- RDAscoresSpecies
df1$species <- rownames(RDAscoresSpecies)

df2 <- traitsBiotopeForest %>% 
  select(species, forestAssociation)

df <- merge(df1, df2, by = "species", all.x = TRUE)

# Schritt 2: Gruppierung nach short label für Statistik
df <- df %>%
  group_by(species, forestAssociation) %>%
  summarise(RDA1 = mean(RDA1), .groups = "drop") %>%
  mutate(forestAssociation_short = forestAssociation)  # für Statistik separat merken

# Schritt 3: Sample sizes
n_labels <- df %>%
  count(forestAssociation_short, name = "n") %>%
  mutate(label = paste0(forestAssociation_short, "\n(n = ", n, ")"))

# Mapping zu Labels (für Plot)
df <- left_join(df, n_labels, by = "forestAssociation_short")
df$forestAssociation <- factor(df$label,
                               levels = n_labels$label)

# Schritt 4: Kruskal-Wallis & Dunn
kruskal_result <- kruskal.test(RDA1 ~ forestAssociation_short, data = df)
print(kruskal_result)

dunn_result <- dunnTest(RDA1 ~ forestAssociation_short, data = df, method = "bh")
p_table <- dunn_result$res

# Matrix für multcompLetters
pair_list <- lapply(strsplit(p_table$Comparison, "-"), trimws)
groups <- unique(unlist(pair_list))

p_matrix <- matrix(1, nrow = length(groups), ncol = length(groups),
                   dimnames = list(groups, groups))

for (i in seq_along(pair_list)) {
  pair <- pair_list[[i]]
  pval <- p_table$P.adj[i]
  p_matrix[pair[1], pair[2]] <- pval
  p_matrix[pair[2], pair[1]] <- pval
}

letters_raw <- multcompLetters(p_matrix)$Letters

# DataFrame mit Zuordnung
sig_letters <- data.frame(
  forestAssociation_short = names(letters_raw),
  letters = letters_raw,
  stringsAsFactors = FALSE
)

# Mapping zu gelabelten Plotnamen
sig_letters <- left_join(sig_letters, n_labels, by = "forestAssociation_short")

# Y-Positionen für Plot
sig_letters <- sig_letters %>%
  mutate(pos = 0.44)

# Ergebnis:
# forestAssociation_short = original group (e.g. "1.1")
# label = full label with n = ...
# letters = a, ab, etc.
# pos = position for geom_text()
```


> print(kruskal_result)

	Kruskal-Wallis rank sum test

data:  RDA1 by forestAssociation
Kruskal-Wallis chi-squared = 77.21, df = 4, p-value = 6.79e-16

> dunn_result$res
                         Comparison          Z      P.unadj        P.adj
1      1.1\n(n = 40) - 1.2\n(n = 5)  0.2119101 8.321772e-01 8.321772e-01
2     1.1\n(n = 40) - 2.1\n(n = 83) -3.4232594 6.187499e-04 1.237500e-03
3      1.2\n(n = 5) - 2.1\n(n = 83) -1.6491752 9.911175e-02 1.238897e-01
4     1.1\n(n = 40) - 2.2\n(n = 92) -7.9430742 1.972309e-15 1.972309e-14
5      1.2\n(n = 5) - 2.2\n(n = 92) -3.4949013 4.742371e-04 1.185593e-03
6     2.1\n(n = 83) - 2.2\n(n = 92) -5.5847513 2.340348e-08 1.170174e-07
7  1.1\n(n = 40) - Others\n(n = 16) -4.0350369 5.459371e-05 1.819790e-04
8   1.2\n(n = 5) - Others\n(n = 16) -2.5258203 1.154285e-02 1.923809e-02
9  2.1\n(n = 83) - Others\n(n = 16) -1.9582615 5.019933e-02 7.171333e-02
10 2.2\n(n = 92) - Others\n(n = 16)  1.1473464 2.512385e-01 2.791539e-01



### Plot
```{r}
gRDA_forest = ggplot(df, aes(x = forestAssociation, y = RDA1)) +
  geom_boxplot(aes(fill = forestAssociation), color = "black") +
  stat_summary(fun = mean, geom = "point", shape = 20, size = 3, color = "red") +
  geom_text(data = sig_letters,
            aes(x = label, y = pos, label = letters),
            inherit.aes = FALSE, size = 4) +
  labs(x = "Forest association", 
       y = "RDA1 - Species score (Reverse of encroachment)") +
  scale_y_continuous(limits = c(-0.5, 0.47)) +
  scale_fill_manual(values = c(
    "#f0f0f0",  # very light grey
    "#d9d9d9",  # light grey
    "#bdbdbd",  # medium grey
    "#969696",  # darkish grey
    "#525252"   # dark grey
  )) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(),
    panel.background = element_blank(),
    axis.line = element_line(color = "black"),
    legend.position = "none"
  )

gRDA_forest
```






# Biotope
table 33 grasslands in intermediate to heigh elevations
table 34 grasslands in low to intermediate  elevations
table 35 dry grasslands in low to intermediate  elevations

### Data preparation
```{r}
# Schritt 1: RDA-Scores vorbereiten
df1 <- RDAscoresSpecies
df1$species <- rownames(RDAscoresSpecies)

# Schritt 2: Biotop-Zuordnung transformieren (long format)
traits_selected <- traitsBiotopeForest %>%
  dplyr::select(species, table33, tables34_35, Others) %>%
  distinct()

df2 <- traits_selected %>%
  pivot_longer(
    cols = -species,
    names_to = "Biotope_short",
    values_to = "occ"
  ) %>%
  filter(occ > 0)

# Faktorlevels setzen
df2$Biotope_short <- factor(df2$Biotope_short,
                            levels = c("table33", "tables34_35", "Others"))

# Schritt 3: Daten zusammenführen
df <- merge(df1, df2, by = "species", all.x = TRUE)

# Schritt 4: Sample-Size-Labels erstellen
n_labels <- df %>%
  group_by(Biotope_short) %>%
  summarise(n = n(), .groups = "drop") %>%
  mutate(label = paste0(Biotope_short, "\n(n = ", n, ")"))

# Neue gelabelte Faktorvariable
label_map <- n_labels %>%
  rename(Biotope = label)

df <- left_join(df, label_map, by = "Biotope_short")

df$Biotope <- factor(df$Biotope, levels = label_map$Biotope)

# Schritt 5: Kruskal–Wallis-Test
kruskal_result <- kruskal.test(RDA1 ~ Biotope_short, data = df)
print(kruskal_result)

# Schritt 6: Dunn-Test mit BH-Korrektur
dunn_result <- dunnTest(RDA1 ~ Biotope_short, data = df, method = "bh")
print(dunn_result)

p_table <- dunn_result$res

pair_list <- strsplit(p_table$Comparison, "-")
pair_list <- lapply(pair_list, trimws)  # führende/trailing Whitespaces weg

# Liste aller vorkommenden Gruppen (bereinigt)
groups <- unique(unlist(pair_list))

# Matrix mit Namen vorbereiten
p_matrix <- matrix(1, nrow = length(groups), ncol = length(groups),
                   dimnames = list(groups, groups))

# Matrix füllen mit adjustierten p-Werten
for (i in seq_along(pair_list)) {
  pair <- pair_list[[i]]
  pval <- p_table$P.adj[i]
  p_matrix[pair[1], pair[2]] <- pval
  p_matrix[pair[2], pair[1]] <- pval
}

# Letzter Check: rownames/colnames
rownames(p_matrix) <- colnames(p_matrix) <- trimws(rownames(p_matrix))

# Buchstaben berechnen
letters <- multcompLetters(p_matrix)$Letters

# Ausgabe in DataFrame (jetzt eindeutig)
sig_letters <- data.frame(
  Biotope_short = names(letters),
  letters = letters,
  stringsAsFactors = FALSE
)

# Schritt 9: Label-Zuordnung & Y-Position für Plot
sig_letters <- left_join(sig_letters, label_map, by = "Biotope_short")

positions <- df %>%
  distinct(Biotope) %>%
  mutate(pos = 0.48)

sig_letters <- left_join(sig_letters, positions, by = "Biotope")

# Ergebnis: sig_letters enthält:
# - Biotope (mit n-label)
# - zugehörige Buchstaben (z. B. "a", "ab", "b")
# - y-Position für geom_text()
```


> print(kruskal_result)

	Kruskal-Wallis rank sum test

data:  RDA1 by Biotope_short
Kruskal-Wallis chi-squared = 40.704, df = 2, p-value = 1.45e-09


> print(dunn_result)
Dunn (1964) Kruskal-Wallis multiple comparison
  p-values adjusted with the Benjamini-Hochberg method.

             Comparison         Z      P.unadj        P.adj
1      Others - table33 -4.144239 3.409434e-05 5.114152e-05
2  Others - tables34_35 -5.797495 6.731277e-09 2.019383e-08
3 table33 - tables34_35 -0.359283 7.193834e-01 7.193834e-01

### Plot
```{r}
biotope_labels <- c(
  "table33\n(n = 38)"       = "Table 33\n(n = 38)",
  "tables34_35\n(n = 75)"   = "Tables 34/35\n(n = 75)",
  "Others\n(n = 140)"       = "Other biotopes\n(n = 140)"
)

gRDA_biotope = ggplot(df, aes(x = Biotope, y = RDA1)) +
  geom_boxplot(aes(fill = Biotope), color = "black") +
  stat_summary(fun = mean, geom = "point", shape = 20, size = 3, color = "red") +
  geom_text(data = sig_letters,
            aes(x = Biotope, y = pos, label = letters),
            inherit.aes = FALSE, size = 4) +
  labs(x = "Species associated to tables 33-35 §30 biotopes", 
       y = "RDA1 - Species score (Reverse of encroachment)") +
  scale_y_continuous(limits = c(-0.5, 0.5)) +
   scale_x_discrete(labels = biotope_labels) +  # <- This line does the relabeling
  scale_fill_manual(values = c("#d9d9d9", "#969696", "#525252")) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(),
    panel.background = element_blank(),
    axis.line = element_line(color = "black"),
    legend.position = "none"  # optional: hide legend
  ); gRDA_biotope
```



Tafel 33: Alpine Rase
34 & 35 zusammengefasst: Grünland mäßig

frischer Standorte, Magerstandorte

Bewertungindex: bei 1, schon biotope als gesichert

hypothese: bei früher beweiundg verschiebt sich mehr richtung 34/35





```{r}
save(
  gRDA_forage,gRDA_biotope, gRDA_forest,
  file="results/RDA_plots.RData")
```
