==> r-code/05_3_ordinations - Classification with Isopam.Rmd <==
---
title: "Ordination analyses"
author: "Bernd Panassiti"
date: 'created: 28.11.2024, last modified: `r format(Sys.Date(), format="%d.%m.%Y")`'
output: pdf_document
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
library(knitr)
opts_knit$set(root.dir='../')                     # definining working directory; or normalizePath('../')
opts_chunk$set(
  fig.align='center',# aligns all figures
  echo=TRUE,     # shows r-code
  message=FALSE, # suppresses library outputs
  warnings=FALSE,# suppresses library outputs
  tidy=TRUE,  # prevents the source code from running off a pdf page
  dev='pdf')
```


# load packages
```{r}
#Install/load pacman
if(!require(pacman)){install.packages("pacman");require(pacman)}
#Install packages
p_load(tidyverse,dplyr,vegan,ggplot2,FSA, ggpubr)
```

# load data and functions
```{r}
load("data/workingdata.RData")

source("r-code/00_functions.r")
```





# Classifikation mit Isopam
## Run Isopam
```{r}
rm(.Random.seed, envir=globalenv())
set.seed(2025)

ip <- isopam::isopam(veg_layer_red)
```


## Hierachie-Style nehmen für gleiche Benennung wie Isotab
```{r Isopam}
clusters <- isopam::clusters(ip, 3, style = "hier") 
clusters
clusters_df <- as.data.frame(clusters)
clusters_df <- clusters_df %>%
                rownames_to_column("Plot")
plot(ip)
```


## How many observations per cluster?
```{r}
ip3 <- ip$hier$lev.3


table(ip3)
```




## Plant species associated with clusters

### Isotab 3 - Mit 12 Clustern
```{r}
it3 <- isopam::isotab(ip, 3)
it3

#Tabelle aus Isotab machen
it3_df <- as.data.frame(it3$tab)
it3_df <- it3_df %>%
  rownames_to_column("species")


plot(it3, labels = TRUE)
```




# Boxplot with encroachment and clusters
## Prepare data
```{r}
df1 = plot_data %>% dplyr::rename(site_id =Alm,
                                 encroachment = Cover_Juvenile)

df2 = ip$hier
df2$site_id = rownames(df2)

df = merge(df1,df2,by="site_id", all.x=TRUE)

df$cluster_ip3 <- as.factor(ip3)

df %<>% dplyr::select(site_id,encroachment,cluster_ip3)
```



## Kruskal–Wallis + Dunn test
```{r}
rm(.Random.seed, envir=globalenv())
set.seed(2025)

# Kruskal–Wallis test
kruskal.test(encroachment ~ cluster_ip3, data = df)

# Dunn's post-hoc test (with Holm correction)
dunn_result <- FSA::dunnTest(encroachment ~ cluster_ip3, data = df, method = "holm")
dunn_result
dunn_df <- dunn_result$res
```

> kruskal.test(encroachment ~ cluster_ip3, data = df)

	Kruskal-Wallis rank sum test

data:  encroachment by cluster_ip3
Kruskal-Wallis chi-squared = 38.692, df = 11, p-value = 5.976e-05


### Keep only significant comparisons
```{r}
sig_dunn <- subset(dunn_df, P.adj < 0.05)

# Convert to list of pairs for ggpubr
comparisons_sig <- strsplit(as.character(sig_dunn$Comparison), " - ")
```


## Plot
```{r, warning=FALSE}
n_labels <- df %>%
  group_by(cluster_ip3) %>%
  summarise(n = n()) %>%
  mutate(label = paste0(cluster_ip3, "\n(n=", n, ")"))

isopamBoxplot = ggplot(df, aes(x = cluster_ip3, y = encroachment, fill = cluster_ip3)) +
  geom_boxplot() +
  stat_summary(fun = mean, geom = "point", shape = 20, size = 4, color = "red") +
  scale_fill_manual(values = rev(gray.colors(12))) +
  stat_compare_means(comparisons = comparisons_sig,
                     method = "wilcox.test",  # same as Dunn pairwise
                     label = "p.signif",
                     hide.ns = TRUE) +        # only show sig
  labs(x = "Isopam clusters", 
       y = "Woody encroachment (%)", 
       fill = "Cluster") +
  theme_minimal() +
  theme(axis.text.x = element_text(size = 10))+
  scale_x_discrete(labels = n_labels$label);isopamBoxplot
```




# NMDS mit Isopam
## Run NMDS
```{r NMDS Isopam}
rm(.Random.seed, envir=globalenv())
set.seed(2025)

nmdsIsopam <- vegan::metaMDS(veg_layer_red,distance = "bray",  k = 2, trymax = 100)
autoplot(nmdsIsopam, obscolor = ip3)
```





## Prepare NMDS data
```{r}
 # 1) Extrahiere die NMDS-Punkte
nmds_data <- as.data.frame(nmdsIsopam$points) 

# 2) Füge die Verbuschung-Information hinzu
nmds_data$Verbuschung <- as.factor(plot_data$Verbuschung)  # 

# 3) Nur ausführen, wenn die Verbuschung mit 1 und 0 in plot_data ist!
nmds_data$Verbuschung = ifelse(nmds_data$Verbuschung == 0,
                               "unverbuscht", "verbuscht") 
nmds_data$Cluster <- as.factor(clusters_df$clusters)
```


## Plot NMDS
```{r}
rm(.Random.seed, envir=globalenv())
set.seed(2025)

# arrows
plot_data_sorted <- plot_data[match(rownames(veg_layer_red), plot_data$Plot_ID), ]

env_fit <- envfit(nmdsIsopam,
              plot_data_sorted[, c(
                "Elevation","Cover_Juvenile", "Hangneigung")],
              permutations = 999)

vector_fit <- as.data.frame(env_fit$vectors$arrows)
vector_fit$Variable <- rownames(vector_fit)
vector_fit <- vector_fit %>% rename(MDS1 = NMDS1)
vector_fit <- vector_fit %>% rename(MDS2 = NMDS2)

vector_fit$label_x <- vector_fit$MDS1
vector_fit$label_y <- vector_fit$MDS2 + 0.1

# fin tuning
vector_fit[vector_fit$Variable=="Elevation","label_x"] = -0.8
vector_fit[vector_fit$Variable=="Elevation","label_y"] = -0.6
vector_fit[vector_fit$Variable=="Hangneigung","label_y"] = 0.97


vector_fit <- vector_fit %>%
  mutate(Variable = recode(Variable, 
                           "Elevation" = "elevation", 
                           "Cover_Juvenile" = "encroachment",
                           "Hangneigung" = "slope"))



# nmds data
df = nmds_data
df <- df %>% rename(treatment = Verbuschung,
                    cluster = Cluster)

df <- df %>%
  mutate(treatment = recode(treatment, 
                           "unverbuscht" = "control", 
                           "verbuscht" = "encroachment"))
df$treatmeant <- as.factor(df$treatment)

isopamNMDS = ggplot() +
  geom_point(
    data = df,
    aes(x = MDS1, y = MDS2, color = cluster, shape = treatment),
    size = 3) +
  labs(
    title = "",
    x = "NMDS1",
    y = "NMDS2"
  ) +
  scale_color_manual(values = rev(gray.colors(12))) +
  geom_segment(data = vector_fit, aes(x = 0, y = 0, 
                                      xend = MDS1, yend = MDS2),
               arrow = arrow(length = unit(0.2, "cm")), 
               color = "black", linewidth = 1) +
  geom_text(data = vector_fit, aes(x = label_x, y = label_y, label = Variable),
                        color = "black", vjust = 0.1) +
  theme_minimal()+
  annotate("text", x = Inf, y = -Inf,
           label = paste("Stress =", nmdsIsopam$stress),
           hjust = 1.1, vjust = -0.5, size = 4);isopamNMDS
```


## Summary
```{r}
rm(.Random.seed, envir=globalenv())
set.seed(2025)

# Basic summary
nmdsIsopam

# Stress value
nmdsIsopam$stress

# Scores (site and species)
scores(nmdsIsopam)

# Environmental fit (optional, if you have env variables)
# Assume env_data contains e.g., elevation, slope, etc.
env_fit
```



> nmdsIsopam

Call:
vegan::metaMDS(comm = veg_layer_red, distance = "bray", k = 2,      trymax = 100) 

global Multidimensional Scaling using monoMDS

Data:     wisconsin(sqrt(veg_layer_red)) 
Distance: bray 

Dimensions: 2 
Stress:     0.2050116 
Stress type 1, weak ties
Best solution was repeated 1 time in 20 tries
The best solution was from try 13 (random start)
Scaling: centring, PC rotation, halfchange scaling 
Species: expanded scores based on ‘wisconsin(sqrt(veg_layer_red))


> nmdsIsopam$stress
[1] 0.2050116


Stress between 0.1 and 0.2 is usually considered moderate but still interpretable; stress > 0.2 should be used with caution, so it’s good to explicitly note this.


> env_fit

***VECTORS

                  NMDS1    NMDS2     r2 Pr(>r)    
Elevation      -0.89832 -0.43935 0.3259  0.001 ***
Cover_Juvenile  0.99294 -0.11858 0.0904  0.002 ** 
Hangneigung    -0.39451  0.91889 0.1893  0.001 ***
---
Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
Permutation: free
Number of permutations: 999





# Ellenberg
```{r}
plot_data_sorted <- plot_data[match(rownames(veg_layer_red), plot_data$Plot_ID), ]

fit <- envfit(nmdsIsopam, zeigerwerte_mittel[, c("Lichtzahl", "Temperaturzahl", "Feuchtezahl", "Reaktionszahl", "Naehrstoffzahl")], permutations = 999)
vector_fit <- as.data.frame(fit$vectors$arrows)
vector_fit$Variable <- rownames(vector_fit)
vector_fit <- vector_fit %>% rename(MDS1 = NMDS1)
vector_fit <- vector_fit %>% rename(MDS2 = NMDS2)

# ggplot2-Plot
vector_fit$label_x <- vector_fit$MDS1 + 0.19

nmds_data$Verbuschung <- as.factor(nmds_data$Verbuschung)

isopamNMDS = ggplot() +
  geom_point(
    data = nmds_data,
    aes(x = MDS1, y = MDS2, color = Cluster, shape = Verbuschung),
    size = 3
  ) +
  labs(
    title = "NMDS mit Isopam-Klassifikation und Verbuschung",
    x = "NMDS1",
    y = "NMDS2"
  ) +
  scale_color_manual(values = rev(gray.colors(12))) +
  geom_segment(data = vector_fit, aes(x = 0, y = 0, xend = MDS1, yend = MDS2),
               arrow = arrow(length = unit(0.2, "cm")), color = "black", linewidth = 1) +
  geom_text(data = vector_fit, aes(x = MDS1, y = MDS2, label = Variable),
                        color = "black", vjust = 0.1) +
  theme_minimal()+
  annotate("text", x = Inf, y = -Inf,
           label = paste("Stress =", nmdsIsopam$stress),
           hjust = 1.1, vjust = -0.5, size = 4);isopamNMDS
```


# save
```{r}
save(
  ip, # isopam classification
  it3, # isotab, How many observations per cluster?
  nmdsIsopam,
  isopamBoxplot,
  isopamNMDS,
  file="results/isopam.RData"
)
```
