==> r-code/07_final_plots_for_publication.Rmd <==
---
title: "Final plots for publication"
author: "Bernd Panassiti"
date: 'created: 28.11.2024, last modified: `r format(Sys.Date(), format="%d.%m.%Y")`'
output: pdf_document
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
library(knitr)
opts_knit$set(root.dir='../')                     # definining working directory; or normalizePath('../')
opts_chunk$set(fig.align='center',                # aligns all figures
                echo=TRUE,                        # shows r-code
                message=FALSE,                    # suppresses library outputs
                warnings=FALSE,                   # suppresses library outputs
                tidy=TRUE,  # prevents the source code from running off a pdf page
                dev='pdf')
```

# Load packages
```{r}
#Install/load pacman
if(!require(pacman)){install.packages("pacman");require(pacman)}
#Install packages
p_load(tidyverse,dplyr,ggplot2,ggpubr,rcompanion,grid)
```

# Load working data
```{r}
load("data/workingdata.RData")
load("results/descriptive_analyses-final_plots.RData")
load("results/RDA_plots.RData")
load("results/isopam.RData")
load("results/regression_results.RData")
load("results/ordinations_NMDS.RData")
```

# Main document

## Q1 - Figure 2
```{r}
Fig2

# save plot
size <- dev.size("in")
ggsave(plot = last_plot(), filename =
         paste("figures/Fig2.png",sep=""),
       width = size[1], 
       height = size[2], 
       units = "in",
       dpi = 300,
       bg = "white")
```



## Q1 - Figure 3
forest, biotop, forage
```{r}
combined_plot = ggarrange(
  gRDA_forest, gRDA_biotope, gRDA_forage,
          labels = c("(A) ", "(B)", "(C)"),
          ncol = 3, nrow = 1,
          label.x = c(-0.07, -0.07, -0.07))

# save plot
# Add extra left margin to gRDA_forest
gRDA_forest <- gRDA_forest + 
  theme(plot.margin = margin(t = 5, r = 5, b = 5, l = 20))  # l = 60 pt ≈ ~0.85 inches

# Combine again
combined_plot <- ggarrange(
  gRDA_forest, gRDA_biotope, gRDA_forage,
  labels = c("(A)", "(B)", "(C)"),
  ncol = 3, nrow = 1,
  label.x = c(0, -0.07, -0.07)
);combined_plot

# Save using ggsave
ggsave("figures/Fig3.png", plot = combined_plot,
       width = 15, height = 7, 
       dpi = 300, units = "in", 
       bg = "white")
```

## Q1 - Figure 4
```{r}
empty_plot <- ggplot() + theme_void()

ggarrange(isopamBoxplot,
                  empty_plot,
                  isopamNMDS,
labels = c("(A)", "", "(B)"),
          ncol = 3,
          nrow = 1,
          widths = c(1, 0.1, 1))


ggsave("figures/Fig4.png",
       plot = last_plot(),
       device = "png",
 width = 14, height = 7, 
 units = "in", 
 dpi = 300,
 bg = "white")

```

## Q2 - Figure 5


```{r}
Fig5

ggsave("figures/Fig5.png",
       plot = last_plot(),
       device = "png",
 width = 7, height = 5, 
 units = "in", 
 dpi = 300,
 bg = "white")

```



# Appendix
## S5
```{r}
Appendix_S5

# save plot
size <- dev.size("in")

ggsave(plot = last_plot(), filename =
         paste("figures/Appendix_S5.png",sep=""),
       width = size[1]+2, 
       height = size[2]+2, 
       units = "in",
       dpi = 300,
       bg = "white")
```

## S6
```{r}
Appendix_S6

# save plot
size <- dev.size("in")

ggsave(plot = last_plot(), filename =
         paste("figures/Appendix_S6.png",sep=""),
       width = size[1]+2, 
       height = size[2]+2, 
       units = "in",
       dpi = 300,
       bg = "white")
```
